feat: rudimentary fennelcheck impl
Writes a small fennelcheck script that compiles stdin fennel into lua
and passes to luacheck. For now this shows errors from luacheck in the
fennel file, though in the wrong places and with mangled function names.
Still, the feedback loop is now fast and little things will be caught
earlier.