emacs: switch to corfu
Copy-pastas gagbo's corfu module and updates a few bindings. Works
pretty good! Excited to be done with this pita company-box crash, and
the gdscript completion is already much better (once lsp is connected).

Have not tested with clojure yet, but I'd expect to have to refactor
more of my config, which is also company-based.

Related links:

https://github.com/doomemacs/doomemacs/issues/5600#issuecomment-1072960189
https://git.sr.ht/~gagbo/doom-config/tree/master/item/modules/completion/corfu
https://github.com/doomemacs/doomemacs/issues/5600#issuecomment-1062626198
https://discourse.doomemacs.org/t/new-completion-corfu-module/2685